Switch to postgreSQL instead ov CSV - only /list

+++ server.py
@@ -8,90 +8,35 @@ import util
 # which is subject to change, depending on the server running the software
 
 UPLOAD_FOLDER = '/home/dan/codecool/web/w1/askmate/ask-mate-python/static/img'
-ALLOWED_EXTENSIONS = {'png', 'jpg'}
+UPLOAD_FOLDER_IULIAN = '/home/iulian/PycharmProjects/ask-mate-python/static/img'
+UPLOAD_FOLDER_DAN = '/home/dan/codecool/web/w1/askmate/ask-mate-python/static/img'
+ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
 
 app = Flask(__name__)
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
 
+@app.route('/info')
+def info():
+    return render_template('info.html')
+
+
 @app.route('/', methods=['GET', 'POST'])
-@app.route('/list', methods=['GET', 'POST'])
-def route_list():
-    if request.method == 'GET':
-        sort_by = request.args.get(key='order_by')
+def index():
+    return 'home page'
+
+
+@app.route('/list')
+def list_all_questions():
+    order_by = 'submission_time'
+    order_direction = 'DESC'
+    if request.args.get(key='order_by'):
+        order_by = request.args.get(key='order_by')
+    if request.args.get(key='order_direction'):
         order_direction = request.args.get(key='order_direction')
-        if sort_by == 'submission time' and order_direction == 'desc':
-            questions = data_manager.sort_questions('submission_time', True)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'submission time' and order_direction == 'asc':
-            questions = data_manager.sort_questions('submission_time', False)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'title' and order_direction == 'desc':
-            questions = data_manager.sort_questions('title', True)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'title' and order_direction == 'asc':
-            questions = data_manager.sort_questions('title', False)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'message' and order_direction == 'desc':
-            questions = data_manager.sort_questions('message', True)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'message' and order_direction == 'asc':
-            questions = data_manager.sort_questions('message', False)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'view number' and order_direction == 'desc':
-            questions = data_manager.sort_questions('view_number', True)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'view number' and order_direction == 'asc':
-            questions = data_manager.sort_questions('view_number', False)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'vote number' and order_direction == 'desc':
-            questions = data_manager.sort_questions('vote_number', True)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        elif sort_by == 'vote number' and order_direction == 'asc':
-            questions = data_manager.sort_questions('vote_number', False)
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-        else:
-            questions = data_manager.get_all_questions()
-            return render_template('list.html',
-                                   category=sort_by,
-                                   questions=questions)
-    elif request.method == 'POST':
-        file = request.files['file']
-        new_question_id = util.generate_question_id()
-        if file and allowed_file(file.filename):
-            filename = "q" + str(new_question_id) + secure_filename(file.filename)
-            file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-        new_question = {"id": new_question_id,
-                        "submission_time": data_manager.create_time(),
-                        "view_number": 0,
-                        "vote_number": 0,
-                        'title': request.form['title'],
-                        'message': request.form['message'],
-                        'image': '/static/img/' + filename if file.filename else ''}
-        data_manager.add_question_table(new_question)
-        questions = data_manager.get_all_questions()
-        return render_template('/list.html',
-                               questions=questions)
+    questions = data_manager.get_all_questions(order_by, order_direction)
+    return render_template('list.html',
+                           questions=questions)
 
 
 @app.route('/question/<question_id>/new-answer', methods=['GET'])
@@ -110,15 +55,27 @@ def display_question(question_id):
                                answers=answers,
                                question_id=question_id)
     elif request.method == 'POST':
-        file = request.files['file']
-        new_answer_id = util.generate_answer_id()
-        if file and allowed_file(file.filename):
-            filename = 'a' + str(new_answer_id) + secure_filename(file.filename)
-            file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-        filename = '/static/img/' + filename
-        data_manager.add_answer_to_file(request.form['post_answer'], question_id, new_answer_id, filename)
         question = data_manager.get_question(question_id)
         answers = data_manager.get_answers(question_id)
+        file = request.files['file']
+        new_answer_id = util.generate_answer_id()
+        new_answer = {"id": new_answer_id,
+                      "submission_time": data_manager.create_time(),
+                      "vote_number": 0,
+                      "question_id": question_id,
+                      'message': request.form['message'],
+                      'image': ''}
+        if file and allowed_file(file.filename):
+            filename = "a" + str(new_answer_id) + secure_filename(file.filename)
+            file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+            new_answer = {"id": new_answer_id,
+                            "submission_time": data_manager.create_time(),
+                            "vote_number": 0,
+                            "question_id": question_id,
+                            'message': request.form['message'],
+                            'image': '/static/img/' + filename if file.filename else ''}
+        answers.append(new_answer)
+        data_manager.write_all_answers(answers)
         return render_template('question.html',
                                question=question,
                                answers=answers,
@@ -140,12 +97,22 @@ def delete_question(question_id):
     j = 0
     while j < len(answers):
         if answers[j]['question_id'] == question_id:
+            image_exists = answers[j]['image']
             answers.pop(j)
-        j += 1
+            j = 0
+            # TODO: remove files from answers when teh question is deleted
+            # if image_exists != '':
+            #     try:
+            #         complete_path = f"/home/iulian/PycharmProjects/ask-mate-python{image_name}"
+            #         os.remove(complete_path)
+            #     except IsADirectoryError:
+            #         print("Tried to delete a picture that doesn't exist.")
+            #         continue
+        else:
+            j += 1
     data_manager.write_all_answers(answers)
     questions = data_manager.get_all_questions()
-    complete_path = f"/home/dan/codecool/web/w1/askmate/ask-mate-python{image_name}"
-    os.remove(complete_path)
+
     return render_template('list.html',
                            questions=questions)
 
@@ -224,8 +191,9 @@ def delete_answer(answer_id):
             answers.pop(i)
         i += 1
     data_manager.write_all_answers(answers)
-    complete_path = f"/home/dan/codecool/web/w1/askmate/ask-mate-python{image_name}"
-    os.remove(complete_path)
+    if image_name != '':
+        complete_path = f"/home/dan/codecool/web/w1/askmate/ask-mate-python{image_name}"
+        os.remove(complete_path)
     question = data_manager.get_question(question_id)
     answers = data_manager.get_answers(question_id)
     return render_template('question.html',
@@ -313,5 +281,6 @@ def uploaded_file(filename):
 
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run(debug=True,
+            host='0.0.0.0')
